Move IDifficultyCalculator and IRewardsCalculator to modMcf Use IPowBlockHeader in IDifficultyCalculator Introduce IPoWChainCfg to separete PoW ChainCfg responsabilities
