+++ modAionImpl/src/org/aion/zero/impl/valid/AionDifficultyRule.java
@@ -29,8 +29,9 @@ import static org.aion.base.util.BIUtil.isEqual;
 import java.math.BigInteger;
 import java.util.List;
 
-import org.aion.zero.blockchain.IChainCfg;
-import org.aion.zero.core.IDifficultyCalculator;
+import org.aion.mcf.blockchain.IChainCfg;
+import org.aion.mcf.blockchain.IPoWChainCfg;
+import org.aion.mcf.core.IDifficultyCalculator;
 import org.aion.mcf.valid.GrandParentDependantBlockHeaderRule;
 import org.aion.zero.types.A0BlockHeader;
 import org.aion.zero.types.AionTransaction;
@@ -43,7 +44,7 @@ public class AionDifficultyRule extends GrandParentDependantBlockHeaderRule<A0Bl
 
     IDifficultyCalculator diffCalc;
 
-    public AionDifficultyRule(IChainCfg<IAionBlock, AionTransaction> configuration) {
+    public AionDifficultyRule(IPoWChainCfg<IAionBlock, AionTransaction> configuration) {
         this.diffCalc = configuration.getDifficultyCalculator();
     }
