+++ modMcf/src/org/aion/mcf/blockchain/IChainCfg.java
@@ -22,13 +22,11 @@
  *
  ******************************************************************************/
 
-package org.aion.zero.blockchain;
+package org.aion.mcf.blockchain;
 
 import org.aion.base.type.IBlock;
 import org.aion.base.type.ITransaction;
 import org.aion.mcf.blockchain.IBlockConstants;
-import org.aion.zero.core.IDifficultyCalculator;
-import org.aion.zero.core.IRewardsCalculator;
 import org.aion.mcf.valid.BlockHeaderValidator;
 import org.aion.mcf.valid.ParentBlockHeaderValidator;
 
@@ -42,10 +40,6 @@ public interface IChainCfg<Blk extends IBlock<?, ?>, Tx extends ITransaction> {
     IBlockConstants getConstants();
 
     IBlockConstants getCommonConstants();
-
-    IDifficultyCalculator getDifficultyCalculator();
-
-    IRewardsCalculator getRewardsCalculator();
 
     BlockHeaderValidator createBlockHeaderValidator();
